fix: add sybase 12.5 LastInsertedId() support
adds (some) LastInsertedId() support for the sybase 12.5 compatibility mode. statusRowSybase125 will now be used in the executed sql. it will cast the returned values to int64s that can be cast from in mssql_stmt.go

Support is not perfect because LastInsertedId() returns the last inserted id from the whole connection, not just the scope of the sql statements.